requirements.txt
pytablewriter==0.58.0
PyGithub==1.53
